`l1/maintenance`: extract shared object_sink
Extract the common L1 object writing logic from
`compaction_sink` into a reusable `object_sink` base class.
This deduplicates multipart upload, batch buffering, and
commit logic so that both compaction and leveling sinks can
share it.
