+++ src/v/cloud_topics/level_one/maintenance/compaction/tests/reducer_test.cc
@@ -126,8 +126,7 @@ ss::future<> do_compact(
       io,
       metastore,
       as,
-      config::mock_binding<size_t>(128_MiB),
-      16_MiB);
+      config::mock_binding<size_t>(128_MiB));
     auto reducer = compaction::sliding_window_reducer(
       std::move(src), std::move(sink));
 
@@ -177,8 +176,7 @@ ss::future<> do_compact_with_throwing_sink(
       io,
       metastore,
       as,
-      config::mock_binding<size_t>(128_MiB),
-      16_MiB);
+      config::mock_binding<size_t>(128_MiB));
     auto sink = std::make_unique<l1::throwing_compaction_sink>(
       std::move(inner_sink), std::move(should_roll), std::move(should_throw));
     auto reducer = compaction::sliding_window_reducer(
